+++ src/grscheller/fp/iterables.py
@@ -27,6 +27,7 @@ from enum import auto, Enum
 from typing import Callable, cast, Final, Iterator, Iterable
 from typing import overload, Optional, Reversible, TypeVar
 from .nada import Nada, nada
+from .woException import MB
 
 __all__ = [ 'concat', 'merge', 'exhaust', 'FM',
             'accumulate', 'foldL', 'foldR', 'foldL_sc' ]
@@ -40,6 +41,7 @@ D = TypeVar('D')      # D for data
 L = TypeVar('L')      # L for left
 R = TypeVar('R')      # R for right
 S = TypeVar('S')      # S for sentinel (also used for default values)
+T = TypeVar('T')      # T for sTate
 
 ## Iterate over multiple Iterables
 
@@ -152,41 +154,27 @@ def accumulate(iterable: Iterable[D], f: Callable[[L, D], L],
                 acc = f(acc, ii)
             yield acc
 
-@overload
-def foldL(iterable: Iterable[D], f: Callable[[L, D], L], initial: Optional[L], default: S) -> L|S:
-    ...
-@overload
-def foldL(iterable: Iterable[D], f: Callable[[D, D], D]) -> D|Nada:
-    ...
-@overload
-def foldL(iterable: Iterable[D], f: Callable[[L, D], L], initial: L) -> L:
-    ...
-
 def foldL(iterable: Iterable[D], f: Callable[[L, D], L],
-          initial: Optional[L]=None, default: S|Nada=nada) -> L|S|Nada:
+          initial: Optional[L]=None, default: Optional[S]=None) -> L|S:
     """
     #### Folds iterable left with optional initial value.
 
     * note that ~S can be the same type as ~L
-    * note that when an initial value is not given then ~L = ~D
-    * if iterable empty & no initial value given, return default
+      * note that when an initial value not given then ~L = ~D
+      * if iterable empty & no initial value given, return default
     * traditional FP type order given for function f
-    * raises TypeError if the "iterable" is not iterable
     * never returns if iterable generates an infinite iterator
+    * raises TypeError if the "iterable" is not iterable
 
     """
     acc: L
-    if hasattr(iterable, '__iter__'):
-        it = iter(iterable)
-    else:
-        msg = '"Iterable" is not iterable.'
-        raise TypeError(msg)
+    it = iter(iterable)
 
     if initial is None:
         try:
             acc = cast(L, next(it))  # in this case L = D
         except StopIteration:
-            return cast(S, default)  # if default = nada, then S is Nada
+            return cast(S, default)  # if default = None, then S is None
     else:
         acc = initial
 
@@ -195,40 +183,26 @@ def foldL(iterable: Iterable[D], f: Callable[[L, D], L],
 
     return acc
 
-@overload
-def foldR(iterable: Reversible[D], f: Callable[[D, R], R], initial: Optional[R], default: S) -> R|S:
-    ...
-@overload
-def foldR(iterable: Reversible[D], f: Callable[[D, D], D]) -> D|Nada:
-    ...
-@overload
-def foldR(iterable: Reversible[D], f: Callable[[D, R], R], initial: R) -> R:
-    ...
-
 def foldR(iterable: Reversible[D], f: Callable[[D, R], R],
-          initial: Optional[R]=None, default: S|Nada=nada) -> R|S|Nada:
+          initial: Optional[R]=None, default: Optional[S]=None) -> R|S:
     """
     #### Folds reversible iterable right with an optional initial value.
 
     * note that ~S can be the same type as ~R
-    * note that when an initial value not given then ~R = ~D
-    * if iterable empty & no initial value given, return default
+      * note that when an initial value not given then ~R = ~D
+      * if iterable empty & no initial value given, return default
     * traditional FP type order given for function f
-    * raises TypeError if iterable is not reversible
+    * raises TypeError if "iterable" is not reversible
 
     """
     acc: R
-    if hasattr(iterable, '__reversed__') or hasattr(iterable, '__len__') and hasattr(iterable, '__getitem__'):
-        it = reversed(iterable)
-    else:
-        msg = 'Iterable is not reversible.'
-        raise TypeError(msg)
+    it = reversed(iterable)
 
     if initial is None:
         try:
             acc = cast(R, next(it))  # in this case R = D
         except StopIteration:
-            return cast(S, default)  # if default = nada, then S is nada
+            return cast(S, default)  # if default = None, then S is None
     else:
         acc = initial
 
@@ -237,56 +211,95 @@ def foldR(iterable: Reversible[D], f: Callable[[D, R], R],
 
     return acc
 
-@overload
-def foldL_sc(iterable: Iterable[D], f: Callable[[L, D], L], initial: Optional[L], sentinel: S) -> L|S:
-    ...
-@overload
-def foldL_sc(iterable: Iterable[D], f: Callable[[D, D], D]) -> D|Nada:
-    ...
-@overload
-def foldL_sc(iterable: Iterable[D], f: Callable[[L, D], L], initial: L) -> L:
-    ...
-def foldL_sc(iterable: Iterable[D], f: Callable[[L, D], L|S|Nada],
-             initial: Optional[L]=None, sentinel: S|Nada=nada) -> L|S|Nada:
+def foldL_sc(iterable: Iterable[D|S],
+             f: Callable[[L, D], L|S],
+             initial: Optional[L]=None, sentinel: Optional[S]=None) -> L|S:
     """
     #### Shortcut version of FoldL.
 
-    * if the iterable returns the sentinel value, stop the fold at that point
-    * if f returns the sentinel value, stop the fold at that point
-    * f is never passed the sentinel value
+    * stop fold if sentinel value is encountered
+      * if the iterable returns the sentinel value, stop the fold at that point
+        * f is never passed the sentinel value
+      * if f returns the sentinel value, stop the fold at that point
+        * no dot include sentinel in fold
     * note that ~S can be the same type as ~L
-    * note that when an initial value is not given then ~L = ~D
-    * if iterable empty & no initial value given, return default
+      * note that when an initial value not given then ~L = ~D
+      * if iterable empty & no initial value given, return default
+      * nada: Nada is the "default" default value
     * traditional FP type order given for function f
-    * never returns if iterable infinite & sentinel value not encountered
+    * never returns if iterable infinite and sentinel never encountered
     * raises TypeError if the "iterable" is not iterable
+
     """
     acc: L
-    if hasattr(iterable, '__iter__'):
-        it = iter(iterable)
-    else:
-        msg = '"Iterable" is not iterable.'
-        raise TypeError(msg)
+    it = iter(iterable)
 
     if initial is None:
         try:
             acc = cast(L, next(it))  # in this case L = D
         except StopIteration:
-            return cast(S, sentinel)  # if sentinel = nada, then S is Nada
+            return cast(S, sentinel)  # if sentinel = None, then S is None
     else:
         acc = initial
 
-    for v in it:
-        if v is sentinel or v == sentinel:
+    for d in it:
+        if d is sentinel or d == sentinel:
             break
-        # TODO: Catch exception if f throws one?
-        f_ret = f(acc, v)
+        f_ret = f(acc, cast(D, d))
         if f_ret is sentinel or f_ret == sentinel:
             break
-        else:
-            acc = cast(L, f_ret)  # not the sentinel value
+        acc = cast(L, f_ret)
 
     return acc
 
-# TODO: Shortcut version of foldR
-#       - does not have to be reversible
+def foldR_sc(iterable: Iterable[D|S],
+             f: Callable[[D, R], R],
+             initial: Optional[R]=None, sentinel: Optional[S]=None,
+             pred: Optional[Callable[[D, T], MB[T]]]=None,
+             sTate: Optional[T]=None) -> R|S:
+    """
+    #### Shortcut version of FoldR. TODO: update docstring
+
+    * does not require iterable to be reversible, iterable never "reversed"
+    * stop fold if sentinel value is encountered
+      * if the iterable returns the sentinel value
+        * start the right fold at that point using initial value if given
+        * if an initial value is not given, use previous iterated value
+        * if no previous iterated value, return sentinel
+    * note that ~S can be the same type as ~R
+      * note that when an initial value not given then ~R = ~D
+      * if iterable empty & no initial value given, return default
+      * nada: Nada is the "default" default value
+    * traditional FP type order given for function f
+    * never returns if iterable infinite and sentinel never encountered
+    * raises TypeError if the "iterable" is not iterable
+    * raises RecursionError if maximum recursion depth exceeded
+
+    """
+    it = iter(iterable)
+    if pred is None:
+        pred=lambda d, t: MB(t)
+
+    ds: list[D] = []
+    for d in it:
+        if d is sentinel or d == sentinel:
+            break
+        if (mb := pred(cast(D, d), sTate)):
+            sTate = mb.get()
+            ds.append(cast(D, d))
+        else:
+            break
+
+    if initial is None:
+        if len(ds) == 0:
+            #return sentinel   # TODO: do I need the cast?
+            return cast(S, sentinel)  # if sentinel = None, then S is None
+        else:
+            acc = cast(R, ds.pop())
+    else:
+        acc = initial
+
+    while ds:
+        acc = f(ds.pop(), acc)
+ 
+    return acc